feat: websocket in thread

+++ src/api/routes/candle/candle_service.py
@@ -10,7 +10,7 @@ import json
 
 @dataclass
 class RsiData:
-    rsi: int
+    rsi: float
     start_timestamp: int
 
     def to_json(self):
